Refactor Scene3D smoke runner helper wiring
